Separate mapper related read functions
