+++ speckle_connector/src/constants/geo_constants.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module SpeckleConnector
+  VECTOR_Z = Geom::Vector3d.new(0, 0, 1)
+end
